+++ SpotifyAPI.Web/Models/Response/Image.cs
@@ -2,8 +2,8 @@ namespace SpotifyAPI.Web
 {
   public class Image
   {
-    public int Height { get; set; }
-    public int Width { get; set; }
+    public double Height { get; set; }
+    public double Width { get; set; }
     public string Url { get; set; } = default!;
   }
 }
